+++ expansions/pandemonium/src/main/java/ladysnake/pandemonium/PandemoniumRequiemPlugin.java
@@ -8,7 +8,6 @@ import ladysnake.requiem.api.v1.entity.ability.MobAbilityConfig;
 import ladysnake.requiem.api.v1.entity.ability.MobAbilityRegistry;
 import ladysnake.requiem.api.v1.event.requiem.PossessionStartCallback;
 import ladysnake.requiem.common.entity.ability.MeleeAbility;
-import ladysnake.requiem.common.entity.ability.RangedAttackAbility;
 import net.minecraft.client.MinecraftClient;
 import net.minecraft.client.resource.language.I18n;
 import net.minecraft.entity.EntityType;