+++ src/main/java/ladysnake/requiem/common/gamerule/package-info.java
@@ -1,0 +1,10 @@
+/**
+ * Requiem gamerule definitions
+ */
+@MethodsReturnNonnullByDefault
+@ParametersAreNonnullByDefault
+package ladysnake.requiem.common.gamerule;
+
+import ladysnake.requiem.api.v1.annotation.MethodsReturnNonnullByDefault;
+
+import javax.annotation.ParametersAreNonnullByDefault;